refactor: move status bar logic into separate file
