+++ vscode-cairo/src/config.ts
@@ -3,6 +3,7 @@ import * as vscode from "vscode";
 
 interface ConfigProps {
   enableLanguageServer: boolean;
+  enableStatusBar: boolean;
   languageServerPath: string;
   enableScarb: boolean;
   preferScarbLanguageServer: boolean;
